Chore: Fix vscode intellisense issues
- Fix issues with vscode intellisense. The previous configuration was
  not in sync with conan and cmake.
- Move sorting algorithm tests to their own folder.
