+++ test/array/array_remove.cc
@@ -2,7 +2,7 @@
 #define CAW_TEST_ARRAY_ARRAY_REMOVE_CC_
 
 #include "array_test_setup.cc"
-#include "../../src/array/array.h"
+#include "array/array.h"
 
 class ArrayRemoveTF : public ArrayTF
 {
