+++ test/sorting_algorithms/bubble_sort_unit_test.cc
@@ -1,15 +1,16 @@
-#ifndef CAW_TEST_MERGE_SORT_UNIT_TEST_CC_
-#define CAW_TEST_MERGE_SORT_UNIT_TEST_CC_
-
-#include <memory>
+#ifndef CAW_TEST_BUBBLE_SORT_UNIT_TEST_CC_
+#define CAW_TEST_BUBBLE_SORT_UNIT_TEST_CC_
 
 #include <gtest/gtest.h>
+#include <algorithm>
+#include <memory>
+#include <vector>
 
-#include "../src/sorting_algorithms/merge_sort.h"
+#include "sorting_algorithms/bubble_sort.h"
 
 using IntVector = std::vector<int>;
 
-class MergeSortTestFixture : public ::testing::TestWithParam<IntVector>
+class BubbleSortTestFixture : public ::testing::TestWithParam<IntVector>
 {
 public:
   static const IntVector produceSortedExpected(const IntVector param)
@@ -25,8 +26,8 @@ protected:
 };
 
 INSTANTIATE_TEST_SUITE_P(
-    MergeSort,
-    MergeSortTestFixture,
+    BubbleSort,
+    BubbleSortTestFixture,
     ::testing::Values(
         IntVector{},
         IntVector{1},
@@ -53,14 +54,26 @@ INSTANTIATE_TEST_SUITE_P(
         IntVector{1, 2, 3, 4, -1, 6, 4, 1},
         IntVector{4, 4, 4, 4, 4, 4, 4}));
 
-TEST_P(MergeSortTestFixture, Works)
+TEST_P(BubbleSortTestFixture, Inserts)
 {
   const IntVector param = GetParam();
-  const IntVector expected = MergeSortTestFixture::produceSortedExpected(param);
-  auto mergeSort = std::make_unique<MergeSort<int>>();
-  auto response = mergeSort->sort(param);
+  const IntVector expected = BubbleSortTestFixture::produceSortedExpected(param);
+  const auto bubbleSort = std::make_unique<BubbleSort<int>>();
+  const auto response = bubbleSort->sort(param);
 
+  ASSERT_EQ(response.size(), expected.size());
   ASSERT_EQ(response, expected);
 }
 
-#endif /* CAW_TEST_MERGE_SORT_UNIT_TEST_CC_ */
+TEST_P(BubbleSortTestFixture, LoopsInReverse)
+{
+  const IntVector param = GetParam();
+  const IntVector expected = BubbleSortTestFixture::produceSortedExpected(param);
+  const auto bubbleSort = std::make_unique<BubbleSort<int>>();
+  const auto response = bubbleSort->startFromReverse(param);
+
+  ASSERT_EQ(response.size(), expected.size());
+  ASSERT_EQ(response, expected);
+}
+
+#endif /* CAW_TEST_BUBBLE_SORT_UNIT_TEST_CC_ */
