+++ test/sorting_algorithms/insertion_sort_unit_test.cc
@@ -1,15 +1,15 @@
-#ifndef CAW_TEST_COUNTING_SORT_UNIT_TEST_CC_
-#define CAW_TEST_COUNTING_SORT_UNIT_TEST_CC_
+#ifndef CAW_TEST_INSERTION_SORT_UNIT_TEST_CC_
+#define CAW_TEST_INSERTION_SORT_UNIT_TEST_CC_
 
-#include <algorithm>
 #include <vector>
+#include <algorithm>
 #include <gtest/gtest.h>
 
-#include "../src/sorting_algorithms/counting_sort.h"
+#include "sorting_algorithms/insertion_sort.h"
 
 using IntVector = std::vector<int>;
 
-class CountingSortTestFixture : public ::testing::TestWithParam<IntVector>
+class InsertionSortTestFixture : public ::testing::TestWithParam<IntVector>
 {
 public:
   static const IntVector produceReverseSortedParams(const IntVector params)
@@ -21,12 +21,12 @@ public:
   }
 
 protected:
-  IntVector params;
+  IntVector param;
 };
 
 INSTANTIATE_TEST_SUITE_P(
-    CountingSort,
-    CountingSortTestFixture,
+    InsertionSort,
+    InsertionSortTestFixture,
     ::testing::Values(
         IntVector{3, 2, 1},
         IntVector{3, 2, 1, 6},
@@ -34,14 +34,15 @@ INSTANTIATE_TEST_SUITE_P(
         IntVector{3, 2, 100, 1, 1, 6},
         IntVector{3, 2, 100, 1, 1, 6, -5}));
 
-TEST_P(CountingSortTestFixture, Handles321)
+TEST_P(InsertionSortTestFixture, Handles321)
 {
-  auto instance = new CountingSort();
-  std::vector<int> params = GetParam();
+  auto instance = new InsertionSort();
+  IntVector params = GetParam();
   auto response = instance->loop(params);
-  auto expected = CountingSortTestFixture::produceReverseSortedParams(params);
+  IntVector expected =
+      InsertionSortTestFixture::produceReverseSortedParams(params);
 
   ASSERT_EQ(response, expected);
 }
 
-#endif /* CAW_TEST_COUNTING_SORT_UNIT_TEST_CC_ */
+#endif /* CAW_TEST_INSERTION_SORT_UNIT_TEST_CC_ */
