+++ test/sorting_algorithms/min_heap_sort_unit_test.cc
@@ -1,16 +1,16 @@
-#ifndef CAW_TEST_BUBBLE_SORT_UNIT_TEST_CC_
-#define CAW_TEST_BUBBLE_SORT_UNIT_TEST_CC_
+#ifndef CAW_TEST_MIN_HEAP_SORT_UNIT_TEST_CC_
+#define CAW_TEST_MIN_HEAP_SORT_UNIT_TEST_CC_
 
 #include <gtest/gtest.h>
 #include <algorithm>
 #include <memory>
 #include <vector>
 
-#include "../src/sorting_algorithms/bubble_sort.h"
+#include "sorting_algorithms/min_heap_sort.h"
 
 using IntVector = std::vector<int>;
 
-class BubbleSortTestFixture : public ::testing::TestWithParam<IntVector>
+class MinHeapSortTestFixture : public ::testing::TestWithParam<IntVector>
 {
 public:
   static const IntVector produceSortedExpected(const IntVector param)
@@ -26,8 +26,8 @@ protected:
 };
 
 INSTANTIATE_TEST_SUITE_P(
-    BubbleSort,
-    BubbleSortTestFixture,
+    MinHeapSort,
+    MinHeapSortTestFixture,
     ::testing::Values(
         IntVector{},
         IntVector{1},
@@ -54,26 +54,19 @@ INSTANTIATE_TEST_SUITE_P(
         IntVector{1, 2, 3, 4, -1, 6, 4, 1},
         IntVector{4, 4, 4, 4, 4, 4, 4}));
 
-TEST_P(BubbleSortTestFixture, Inserts)
+TEST_P(MinHeapSortTestFixture, Inserts)
 {
   const IntVector param = GetParam();
-  const IntVector expected = BubbleSortTestFixture::produceSortedExpected(param);
-  const auto bubbleSort = std::make_unique<BubbleSort<int>>();
-  const auto response = bubbleSort->sort(param);
+  const IntVector expected = MinHeapSortTestFixture::produceSortedExpected(param);
+  const auto minHeapSort = std::make_unique<MinHeapSort<int>>();
+  for (auto elem : param)
+  {
+    minHeapSort->insert(elem);
+  }
+  const auto response = minHeapSort->sort();
 
   ASSERT_EQ(response.size(), expected.size());
   ASSERT_EQ(response, expected);
 }
 
-TEST_P(BubbleSortTestFixture, LoopsInReverse)
-{
-  const IntVector param = GetParam();
-  const IntVector expected = BubbleSortTestFixture::produceSortedExpected(param);
-  const auto bubbleSort = std::make_unique<BubbleSort<int>>();
-  const auto response = bubbleSort->startFromReverse(param);
-
-  ASSERT_EQ(response.size(), expected.size());
-  ASSERT_EQ(response, expected);
-}
-
-#endif /* CAW_TEST_BUBBLE_SORT_UNIT_TEST_CC_ */
+#endif /* CAW_TEST_MIN_HEAP_SORT_UNIT_TEST_CC_ */
